Metabolic_sensitivity fixes for sra module
Fixed sra module in metabolic_sensitivity so that it outputs a structure containing all the variables of interest.

Things to still debug:
  * excess_flux variable returns a structure? Need to figure out why
  * zscore normalization outputs error - need to resolve

Added additional script to calculate dynamic range of 6 arrays (dynamic_range.m) and wrote framework of script to
output excel file containing information. This will be supplementary table.